move thermal tests into their own test file test_thermal_speed.py
